\name{lineSummary}
\alias{lineSummary}
\docType{data}
\title{
Sample summary profile data.
}
\description{
This is a small sample object produced by \code{\link{summaryRprof}}
with line profiling enabled.
}
\usage{data(lineSummary)}
\format{
An object produced by  \code{\link{summaryRprof}}.
}

\examples{
data(lineSummary)
lineSummary
}
\keyword{datasets}
